improve create_readable_filenames   - config name as default   - better verbose output   - refactoring

+++ src/taskchain/task/chain.py
@@ -1,6 +1,7 @@
 import abc
 import logging
 import re
+from collections import defaultdict
 from hashlib import sha256
 from itertools import chain
 from pathlib import Path
@@ -371,46 +372,58 @@ class Chain(dict):
                 G.edge(_get_slugname(edge[0]), _get_slugname(edge[1]))
         return G
 
-    def create_readable_filenames(self, name=None, group=None, verbose=True, **kwargs):
-        some_task_found = False
+    def create_readable_filenames(self, groups=None, name=None, verbose=False, keep_existing=False):
+        """
+        Create human readable symlink to data of tasks in the chain.
+        Symlink is in same directory as data, i.e. in directory with all task's data.
+        Name of link is taken from
+            - parameter
+            - task's config, parameter `human_readable_data_name`
+            - name of task's config
+
+        Args:
+            groups (optional): name of group or list of names of groups, for which should be symlinks created
+            name (optional): name of links
+            verbose:
+            keep_existing: rewrite link if link already exists
+        """
+        symlink_actions = defaultdict(list)
+        if type(groups) is str:
+            groups = [groups]
+
         for task_name, task in self.tasks.items():
-            if group is not None and task.group != group:
+            if groups is not None and task.group not in groups:
                 continue
-            else:
-                some_task_found = True
             if not task.has_data:
                 continue
-            self._create_softlink_to_task_data(task, name, verbose=verbose, **kwargs)
-
-        if verbose and not some_task_found:
-            print(f'No task found in group `{group}`')
-
-    def _create_softlink_to_task_data(self, task, name=None, verbose=True, dry=False, keep_existing=False):
-        if name is None:
-            config = task.get_config().get_original_config()
-            name = config.get('human_readable_data_name')
-            if name is None:
-                if verbose:
-                    print(f'no name for task `{task}`')
-                return
-
-        symlink_path = task.path / f'{name}{task.data_path.suffix}'
-        if symlink_path.exists():
-            if keep_existing:
-                action_name = 'keep existing'
-            else:
-                action_name = 'to rewrite' if dry else 'rewriting'
-        else:
-            action_name = 'to create' if dry else 'creating'
+            an, n, sp = self._create_softlink_to_task_data(task, name, keep_existing=keep_existing)
+            symlink_actions[n].append((an, sp))
 
         if verbose:
-            print(f'{action_name:>20}: softlink for task `{task}`: `{symlink_path}`')
-        if not dry:
-            if symlink_path.exists():
-                if keep_existing:
-                    return
-                symlink_path.unlink()
+            for name, actions in symlink_actions.items():
+                print(f'{name}')
+                for action_name, symlink_path in actions:
+                    print(f'   {action_name} link to {symlink_path}')
+
+    def _create_softlink_to_task_data(self, task, name=None, keep_existing=False):
+        if name is None:
+            config = task.get_config().get_original_config()
+            if 'human_readable_data_name' in config:
+                name = config['human_readable_data_name']
+            else:
+                name = config.name
+
+        symlink_path = task.path / f'{name}{task.data_path.suffix}'
+
+        action_name = 'keep existing'
+        if symlink_path.exists() and not keep_existing:
+            symlink_path.unlink()
+            action_name = 'rewriting'
+        if not symlink_path.exists():
             symlink_path.symlink_to(task.data_path.relative_to(symlink_path.parent), task.data_path.is_dir())
+            if action_name != 'rewriting':
+                action_name = 'creating'
+        return action_name, name, symlink_path
 
 
 class TaskParameterConfig(Config):
